refactor: TaskItem に依存したコンポーネントを1つにまとめるため
ディレクトリ構成の見直し
TaskItem 関連をサブディレクトリへ移動
