+++ todo-app/src/components/TaskItem/TaskItem.tsx
@@ -8,7 +8,7 @@ import { CheckIcon, CloseIcon, DeleteIcon, EditIcon } from "@chakra-ui/icons";
 import { Spacer } from "@chakra-ui/layout";
 import { Checkbox } from "@chakra-ui/react";
 import { DraggableStateSnapshot } from "react-beautiful-dnd";
-import { TaskType } from "../types/TaskType";
+import { TaskType } from "../../types/TaskType";
 import TaskItemCustomEditablePreview from "./TaskItemCustomEditablePreview";
 import TaskItemCustomFlex from "./TaskItemCustomFlex";
